+++ examples/arenera.jl
@@ -19,11 +19,15 @@ function testRun(seed::Int64)
     pat_type = jOpqua.newPathogenType(
         "pat_type", possible_alleles="AB",
         mean_mutations_per_replication=0.001,
-        contactCoefficient=s::String->1.0+(0.1*(4.0-hamming(s,"BBBB"))/4.0),
+        contactCoefficient=s::String -> 1.0 + (0.1 * (4.0 - hamming(s, "BBBB")) / 4.0),
+        receiveContactCoefficient=s::String -> 0.0,
     )
 
     res_type = jOpqua.newResponseType("res_type")
-    pop_type = jOpqua.newPopulationType("pop_type")
+    pop_type = jOpqua.newPopulationType(
+        "pop_type",
+        pathogenFractions=jOpqua.pathogenFractionsProportionalFitness,
+    )
 
     num_hosts = 10000
     num_infected = Int(num_hosts * 0.5)
@@ -54,25 +58,25 @@ function testRun(seed::Int64)
     println(output.compartment_vars["pop"][:, end])
     # println(output.host_samples["pop"][:, end][1:3])
 
-    compartment_data = jOpqua.saveCompartments(output, "examples/compartment_test.csv")
-    jOpqua.plotCompartments(compartment_data, ["pop"], "examples/compartment_test.png")
+    # compartment_data = jOpqua.saveCompartments(output, "examples/compartment_test.csv")
+    # jOpqua.plotCompartments(compartment_data, ["pop"], "examples/compartment_test.png")
 
-    his_dat = jOpqua.saveHistory(output, "examples/history_test.csv")
-    composition_data = jOpqua.saveComposition(
-        his_dat, "examples/composition_test.csv",
-        num_top_sequences=7, track_specific_sequences=["AAAA","BBBB"]
-    )
-    jOpqua.plotComposition(composition_data, "examples/composition_test.png", normalized=true)
+    # his_dat = jOpqua.saveHistory(output, "examples/history_test.csv")
+    # composition_data = jOpqua.saveComposition(
+    #     his_dat, "examples/composition_test.csv",
+    #     num_top_sequences=7, track_specific_sequences=["AAAA", "BBBB"]
+    # )
+    # jOpqua.plotComposition(composition_data, "examples/composition_test.png", normalized=true)
 
-    nwks = jOpqua.saveNewick(output, "examples/pathogen_newick_test.nwk")
-    for nwk in nwks
-        jOpqua.plotPhylogeny(nwk, "examples/pathogen_newick_test.png")
-    end
+    # nwks = jOpqua.saveNewick(output, "examples/pathogen_newick_test.nwk")
+    # for nwk in nwks
+    #     jOpqua.plotPhylogeny(nwk, "examples/pathogen_newick_test.png")
+    # end
 end
 
 @time testRun(1)
 
-# @profview testRun(0)
+@profview testRun(0)
 @time testRun(0)
 
 # Result M3 Max 64 GB 9 Feb (second run) seed 0: