+++ src/parameters/parameters.jl
@@ -51,9 +51,9 @@ const DEFAULT_POPULATION_TYPE = PopulationType(
     0.0,
     true,
     0.0,
-    (h_1::String, h_2::String)->true,
-    g::String->1.0, # takes seq argument, returns Float64
-    g::String->1.0, # takes seq argument, returns Float64
+    (h_1::String, h_2::String) -> true,
+    g::String -> 1.0, # takes seq argument, returns Float64
+    g::String -> 1.0, # takes seq argument, returns Float64
     SA[ # order defined in COEFFICIENTS
         0.0, 1.0, 0.0,
         0.0, 1.05, 0.0,